[platform_device_loader] add missing list pki test.
